% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer_manager.R
\name{activate_rflow_viewer}
\alias{activate_rflow_viewer}
\title{Activate Rflow Viewer Protection}
\usage{
activate_rflow_viewer()
}
\description{
Redirects viewer content to browser while Rflow is running
}
\keyword{internal}
