% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{clear_old_sessions}
\alias{clear_old_sessions}
\title{Clear old sessions}
\usage{
clear_old_sessions(days_to_keep = 90)
}
\arguments{
\item{days_to_keep}{Number of days to keep sessions}
}
\description{
Clear old sessions
}
\keyword{internal}
