% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_session}
\alias{create_session}
\title{Create new session}
\usage{
create_session(working_dir = getwd())
}
\arguments{
\item{working_dir}{Current working directory}
}
\value{
Session ID
}
\description{
Create new session
}
\keyword{internal}
