% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer_manager.R
\name{deactivate_rflow_viewer}
\alias{deactivate_rflow_viewer}
\title{Deactivate Rflow Viewer Protection}
\usage{
deactivate_rflow_viewer()
}
\description{
Restores normal viewer behavior
}
\keyword{internal}
