% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_source_search.R
\name{find_r_function}
\alias{find_r_function}
\title{Find R Function Implementation}
\usage{
find_r_function(func_name)
}
\arguments{
\item{func_name}{Name of R function (e.g., "mean", "sum", "lm")}
}
\value{
Information about where the function is implemented
}
\description{
Locate where a specific R function is implemented in the C source code.
}
