% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_sessions}
\alias{get_sessions}
\title{Get all sessions}
\usage{
get_sessions(limit = 50)
}
\arguments{
\item{limit}{Maximum number of sessions to return}
}
\value{
Data frame of sessions
}
\description{
Get all sessions
}
\keyword{internal}
