% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{import_session}
\alias{import_session}
\title{Import session from JSON}
\usage{
import_session(input_file)
}
\arguments{
\item{input_file}{Input file path}
}
\value{
Session ID
}
\description{
Import session from JSON
}
\keyword{internal}
