% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer_manager.R
\name{is_rflow_viewer_active}
\alias{is_rflow_viewer_active}
\title{Check if Rflow Viewer is Active}
\usage{
is_rflow_viewer_active()
}
\value{
Logical indicating if Rflow is protecting the viewer
}
\description{
Check if Rflow Viewer is Active
}
