% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_session_messages}
\alias{load_session_messages}
\title{Load session messages}
\usage{
load_session_messages(session_id)
}
\arguments{
\item{session_id}{Session ID}
}
\value{
Data frame of messages
}
\description{
Load session messages
}
\keyword{internal}
