% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_manager.R
\name{open_folder}
\alias{open_folder}
\title{Set the current working folder for Rflow}
\usage{
open_folder(folder_path = NULL)
}
\arguments{
\item{folder_path}{Path to the folder}
}
\description{
Set the current working folder for Rflow
}
