% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer_manager.R
\name{open_in_browser}
\alias{open_in_browser}
\title{Manually Open Content in Browser}
\usage{
open_in_browser(content)
}
\arguments{
\item{content}{Content to display (htmlwidget, ggplot, etc.)}
}
\description{
Force content to open in browser instead of viewer
}
\examples{
\dontrun{
library(leaflet)
map <- leaflet() \%>\% addTiles()
open_in_browser(map)
}
}
