% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_manager.R
\name{scan_folder}
\alias{scan_folder}
\title{Scan folder structure}
\usage{
scan_folder(folder_path, max_depth = 3)
}
\arguments{
\item{folder_path}{Path to scan}

\item{max_depth}{Maximum depth to scan}
}
\value{
List with folder information
}
\description{
Scan folder structure
}
\keyword{internal}
