% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-system.R
\name{tool_run_command}
\alias{tool_run_command}
\title{System command execution tools
Allows the AI to navigate the file system, create folders, delete files, etc.
Execute system commands}
\usage{
tool_run_command()
}
\description{
System command execution tools
Allows the AI to navigate the file system, create folders, delete files, etc.
Execute system commands
}
