% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{update_session_title}
\alias{update_session_title}
\title{Update session title}
\usage{
update_session_title(session_id, title)
}
\arguments{
\item{session_id}{Session ID}

\item{title}{New title}
}
\description{
Update session title
}
\keyword{internal}
